Chain multiple AttributeValueType nodes, instead of nesting.
